refactor(dim): srsDimension を uint で統一し int キャストを除去
xs:positiveInteger → uint のマッピングに合わせ、derefDim/preferDim/globalDim/
inheritDim 等の次元系変数を uint に統一。ToPoints/PointFromFlat/LineStringFromFlat/
RingFromFlat まで含めて uint ベースに変更し、int への変換を完全に除去。

+++ gml2_1_2/internal/decode_point.go
@@ -26,7 +26,7 @@ func pointFromXML(x *gen.PointType) (core.Point, error) {
 		if err != nil {
 			return core.Point{}, err
 		}
-		return core.PointFromFlat(coords, len(coords))
+		return core.PointFromFlat(coords, uint(len(coords)))
 	}
 	if x.Coord != nil {
 		y := 0.0
